Library bug fixes, ECU bug fixes

+++ HyTech17_Library/MC_firmware_information.cpp
@@ -16,18 +16,18 @@ void MC_firmware_information::load(uint8_t buf[8]) {
   memcpy(&message, buf, sizeof(CAN_message_mc_firmware_information_t));
 }
 
-int16_t MC_firmware_information::get_eeprom_version_project_code() {
+uint16_t MC_firmware_information::get_eeprom_version_project_code() {
   return message.eeprom_version_project_code;
 }
 
-int16_t MC_firmware_information::get_software_version() {
+uint16_t MC_firmware_information::get_software_version() {
   return message.software_version;
 }
 
-uint32_t MC_firmware_information::get_date_code_mmdd() {
+uint16_t MC_firmware_information::get_date_code_mmdd() {
   return message.date_code_mmdd;
 }
 
-uint32_t MC_firmware_information::get_date_code_yyyy() {
+uint16_t MC_firmware_information::get_date_code_yyyy() {
   return message.date_code_yyyy;
 }
